refactor to split MIIServerClient
